Move deleteStore tests out of FDBRecordStoreStateCacheTest
Some of the new additions around delete store, seemed to really be
going beyond the cache, so I moved all the delete store tests out.

This was pretty mechanical, I just cut+pasted the tests, move the
static objects to the utils class, and updated references.

I did not move VersionIndexTest.deleteStoreWithUncommittedVersionData
because that depended on a lot more inner details of the
VersionIndexTest
